requirements.txt
openai
urllib3
feedparser
requests
beautifulsoup4
feedgen
pytz
groq
smolagents[litellm]
dotenv
IProgress
ipywidgets
jupyter
tqdm
markdownify
google-generativeai
google-genai 
wave